Related to #5478
was facing error regarding pom.xml file -lossy-conversions, to fix it needed to make some type casting and cleanup of code in AliquotSum.java, Caesar.java, PerfectNumber.java classes respectively.

now new VolumeOfFrustumOfCone method created and all test cases passed successfully of VolumeTest class

+++ src/main/java/com/thealgorithms/maths/AliquotSum.java
@@ -56,7 +56,7 @@ public final class AliquotSum {
         // if n is a perfect square then its root was added twice in above loop, so subtracting root
         // from sum
         if (root == (int) root) {
-            sum -= root;
+            sum -= (int) root;
         }
         return sum;
     }
